refactor: 페이지들 ts로 변경
